An overhaul of resources as they apply to routing
This fixes the bug whereby routing sometimes made no sense. It turns out
that there was a bug in the resource values which made TAEB really really
scared of walking diagonally, unless it could also do the same thing via
walking orthogonally (which was usually the case, which was why the bug
was hard to find). Risk data is now added to plans to help catch similar
trouble in the future (the spending plan and its numerical value are
given so that they can be compared for sanity reasons).