Refactor tactical routing and add sanity checks
This fixes an obscure bug where TMEs were not getting correctly reset. I
think that's what was causing the Walk to non-adjacent tile bug (which
would happen if trying to route to a non-routable tile on a different level,
which would explain why it hardly ever came up), but I'm not sure.