Don't try failed travels forever
The issue with a fixed backoff upon failing to travel is that in the case
of a routing failure in NetHack's _ command, on some large maps we simply
get stuck doing the same thing again and again. In case we end up changing
our mind, I've added a rather large backoff to travel timeout; it's
possible that a smaller one might work, but I like to err on the safe but
slow side.