Efficiency patch for exploration
Although no semantic changes to exploration are made here, this
precalculates and caches an exploration graph, meaning that instead
of calculating where to explore every turn, we only calculate the
difference between this turn's and last turn's exploration graphs.
This saves about 3 and a half seconds per step compared to the code
before this patch.