Fix threat check for multiturn actions
The effects of this bug meant that TAEB would consider, for instance,
digging into a wall to escape, on the assumption that while it was digging
it would be inside the wall and thus immune to monster attack! This may have
been causing other bugs too; it would be nice if it's causing the back-of-the-
pack ScareMonster bug, although I'm not certain that's the same bug.