Fixing the back-of-the-pack routing bug
A bug that's been plaguing me for months, luckily with a trivial fix.
Although it's true that monsters can't route through other monsters, monsters
can move out of the way; so a square can be vulnerable to attacks by monsters
(in the future) even if there isn't a monster standing on it (now). As a
result, it's a very bad idea to attempt to move onto the square a monster's
on in order to get immunity from enemy attacks, which is what Planar was
trying to do before this fix.