Add chokepoint handling
This isn't actually used for anything yet (apart from pretty debug views),
but eventually it'll be used to optimise tactical routing. Some of the
calculations here are repeated in searching; when I get around to rewriting
the search code, I'll try to use the chokepoint map rather than recalculating
everything.