Pluggable routing algorithms, and a start of the new 'chokepoint' algorithm
The way this works is far too long to fit into a version control comment; I
may make a blog post about it at some point. This adds the possibility for
user-selectable routing algorithms; there are now three, 'world' (ftr),
'level' (the non-ftr algorithm used beforehand), and 'chokepoint' (new).
Each falls back to the previous one in situations where it doesn't apply
(e.g. 'level' falls back to 'world' after changing level, 'chokepoint' falls
back to 'level' if there are threats on the map). The 'chokepoint' algorithm
is very slow in this patch, as I haven't optimised it yet; the idea is that
it can cache parts of routes between turns, even in the face of a changing
map, and hopefully speed up tactical routing massively in the future.