Split Eliminate from PathPast
The code at the moment isn't discerning enough to be able to
distinguish between benefits from killing a monster, and the lesser
benefit from scaring it one step down a corridor so that you can
follow it. Thus, split the plans that generate the two cases, so
that one or the other can be requested.